feat: exists sync

+++ web-server/server/utils/preview.js
@@ -1,10 +1,13 @@
 const shell = require('shelljs');
+const fs = require('fs');
 
 const useStatic = async () => {
   // 删除 preview 静态文件
-  shell.exec('pwd');
-  shell.exec('ls');
-  shell.exec('rm -rf ../../public/preview/*');
+  if (!fs.existsSync('../../public/preview/')) {
+    shell.exec('mkdir -p public/preview');
+  } else {
+    shell.exec('rm -rf ../../public/preview/*');
+  }
   shell.exec('cp -r ./static/* ../../public/preview/');
   process.exit();
 };
